Fix csproj Improve for CI/CD
