+++ ConsoleApp1/Program.cs
@@ -1,105 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics.CodeAnalysis;
 using Tinyhand;
-using Tinyhand.IO;
 
 namespace ConsoleApp1
 {
-    [TinyhandObject]
-    public partial class Callback1_2 : ITinyhandSerializationCallback
-    {
-        [Key(0)]
-        public int X { get; set; }
-
-        [Key(1)]
-        public GenericClass<int> GenericInt { get; set; } = default!;
-
-        [IgnoreMember]
-        public bool CalledBefore { get; private set; }
-
-        [IgnoreMember]
-        public bool CalledAfter { get; private set; }
-
-        [TinyhandObject]
-        internal partial class ChildClass
-        {
-            [Key(0)]
-            public string Name { get; set; } = default!;
-        }
-
-        public Callback1_2(int x)
-        {
-            this.X = x;
-        }
-
-        public Callback1_2()
-        {
-        }
-
-        public void OnBeforeSerialize()
-        {
-            this.CalledBefore = true;
-        }
-
-        void ITinyhandSerializationCallback.OnAfterDeserialize()
-        {
-            this.CalledAfter = true;
-        }
-    }
-
-    [TinyhandObject]
-    public partial class GenericClass<T>
-    {
-        [Key(0)]
-        public T? Member { get; set; }
-    }
-
-    [TinyhandObject]
-    public partial class TestClass
-    {
-        [Key(0)]
-        public int x;
-
-        [Key(2)]
-        public int y;
-
-        [Key(3)]
-        public string?[] stringList = default!;
-
-        [Key(4)]
-        public string?[]? stringList2 = default!;
-
-        [Key(5)]
-        public decimal?[] DecimalArray { get; set; } = { -144m, 456m, null, 78998m, };
-
-        [Key(6)]
-        public double?[] DoubleArray { get; set; } = { -100d, 0d, 123456d, 456789d, null, };
-
-        [Key(7)]
-        public int[,,] z = { { { 0 }, { 1 } }, { { 3 }, { 2 } }, { { 10 }, { 11 } }, { { 13 }, { 12 } } };
-
-        [Key(8)]
-        public KeyValuePair<int, string> kvp = new KeyValuePair<int, string>(3, "test");
-
-        [Key(9)]
-        public Version version = new Version(1, 2);
-
-        [Key(10)]
-        public Lazy<string> ls = new Lazy<string>(() => "test2");
-
-        [Key(11)]
-        public Type type = typeof(int);
-
-        public void Print()
-        {
-            Console.WriteLine(x);
-        }
-    }
-
     [TinyhandObject] // Annote a [TinyhandObject] attribute.
-    public partial class MyClass // partial class is required for source generation.
+    public partial class MyClass // partial class is required for source generator.
     {
         // Key attributes take a serialization index (or string name)
         // The values must be unique and versioning has to be considered as well.
@@ -139,21 +46,14 @@ namespace ConsoleApp1
     {
         static void Main(string[] args)
         {
-            TinyhandModule.Initialize();
-            var b = TinyhandSerializer.Serialize(new Callback1_2());
-            var callBackClass = TinyhandSerializer.Deserialize<Callback1_2>(b);
+            TinyhandModule.Initialize(); // .NET Core 3.1 does not support ModuleInitializerAttribute, so you need to call TinyhandModule.Initialize() before using Tinyhand. Not required for .NET 5.
 
             var myClass = new MyClass() { Age = 10, FirstName = "hoge", LastName = "huga", };
-            b = TinyhandSerializer.Serialize(myClass);
+            var b = TinyhandSerializer.Serialize(myClass);
             var myClass2 = TinyhandSerializer.Deserialize<MyClass>(b);
 
             b = TinyhandSerializer.Serialize(new EmptyClass()); // Empty data
             var myClass3 = TinyhandSerializer.Deserialize<MyClass>(b); // Create an instance and set non-null values of the members.
-
-            var tc = new TestClass() { x = 10 };
-            b = TinyhandSerializer.Serialize(tc);
-            var tc2 = TinyhandSerializer.Deserialize<TestClass>(b);
-            tc2?.Print();
         }
     }
 }
