+++ Tinyhand/Tinyhand.csproj
@@ -2,13 +2,10 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup>
-    <VersionPrefix>0.0.1</VersionPrefix>
     <PackageId>Tinyhand</PackageId>
     <Authors>archi-Doc</Authors>
     <Product>Tinyhand</Product>
