+++ TinyhandGenerator/TinyhandGenerator.csproj
@@ -2,13 +2,10 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
-    <LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup>
-    <VersionPrefix>0.0.1</VersionPrefix>
     <PackageId>TinyhandGenerator</PackageId>
     <Authors>archi-Doc</Authors>
     <Product>TinyhandGenerator</Product>
